refactor: extract popup trigger logic to lib, add PropTypes
Move shouldShowSupportPopup/dismissSupportPopup out of SupportPopup.jsx
into src/lib/supportPopup.js to fix the react-refresh/only-export-components
HMR warning. Add PropTypes declaration for the onClose prop.
