+++ src/components/SupportPopup.jsx
@@ -1,24 +1,5 @@
-import { CURRENT_PATCH } from '../constants';
-
-const STORAGE_KEY = 'bbx-support-popup';
-const SEVEN_DAYS_MS = 7 * 24 * 60 * 60 * 1000;
-
-export function shouldShowSupportPopup() {
-  const raw = localStorage.getItem(STORAGE_KEY);
-  if (!raw) return true;
-  try {
-    const { lastSeen, seenPatch } = JSON.parse(raw);
-    if (seenPatch !== CURRENT_PATCH) return true;
-    if (Date.now() - lastSeen >= SEVEN_DAYS_MS) return true;
-    return false;
-  } catch {
-    return true;
-  }
-}
-
-function dismissSupportPopup() {
-  localStorage.setItem(STORAGE_KEY, JSON.stringify({ lastSeen: Date.now(), seenPatch: CURRENT_PATCH }));
-}
+import PropTypes from 'prop-types';
+import { dismissSupportPopup } from '../lib/supportPopup';
 
 const ExternalLinkIcon = () => (
   <svg className="w-4 h-4 flex-shrink-0" style={{ color: 'var(--color-text-muted)' }} fill="none" stroke="currentColor" viewBox="0 0 24 24">
@@ -26,7 +7,7 @@ const ExternalLinkIcon = () => (
   </svg>
 );
 
-export default function SupportPopup({ onClose }) {
+function SupportPopup({ onClose }) {
   function handleClose() {
     dismissSupportPopup();
     onClose();
@@ -145,3 +126,9 @@ export default function SupportPopup({ onClose }) {
     </div>
   );
 }
+
+SupportPopup.propTypes = {
+  onClose: PropTypes.func.isRequired,
+};
+
+export default SupportPopup;
